Move Animation-related useEffect from RouteVisual to useAnimation within Animation
