+++ src/app/components/routemap/Responsive.tsx
@@ -12,7 +12,8 @@ const Responsive = () => {
       )
     }
   }, [])
-  return { screenWidth }
+
+  return screenWidth
 }
 
 export default Responsive
